requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-extensions==3.2.3
djangorestframework==3.14.0
kombu==5.3.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
pygraphviz==1.11
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
